requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cloudinary==1.24.0
cryptography==3.4.6
defusedxml==0.7.1
dj-config-url==0.1.1
dj-database-url==0.5.0
dj-rest-auth==2.1.4
Django==3.1.7
django-allauth==0.44.0
django-crispy-forms==1.11.1
django-filter==2.4.0
django-heroku==0.3.1
django-multiselectfield==0.1.12
django-rest-auth==0.9.5
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
gunicorn==20.0.4
idna==2.10
Markdown==3.3.4
oauthlib==3.1.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
whitenoise==5.2.0
